\name{JSONXMLUncertWebRTools-package}
\alias{JSONXMLUncertWebRTools-package}
\alias{JSONXMLUncertWeb}
\docType{package}
\title{JSONXMLUncertWeb}
\description{Reads, writes and optionally validates UncertML in JSON/XML}
\details{
\tabular{ll}{
Package: \tab JSONXMLUncertWeb\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-05-29\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Lucy Bastin \email{l.bastin@aston.ac.uk}}
\references{UncertWeb wiki}
\keyword{package, UncertML}
% \usage{Reads, writes and optionally validates UncertML in JSON/XML}
% \seealso{ }
% \examples{ }
